Changed database to local and network to remote. (#47)
We have two sources now local and remote, both together will be exposed via a repository. We will also have a model class that will be used in the app which will be converted from local and remote to make sure app does not know about how local and remote work.